requirements.txt
requests
numpy
pandas
nltk
spacy==2.2
tensorflow==1.14
gensim
pyemd
sklearn
flask
flasgger
xlrd
openpyxl
pdfplumber
bs4
unidecode
flashtext
lxml
bert-serving-server
bert-serving-client
